Added common helper for gating form submit values
