+++ packages/commonwealth/client/scripts/views/pages/Community/Groups/Create/CreateCommunityGroupPage.tsx
@@ -6,8 +6,8 @@ import app from 'state';
 import { useCreateGroupMutation } from 'state/api/groups';
 import Permissions from 'utils/Permissions';
 import { PageNotFound } from '../../../404';
-import { SPECIFICATIONS, TOKENS } from '../../common/constants';
 import { GroupForm } from '../common/GroupForm';
+import { makeGroupDataBaseAPIPayload } from '../common/helpers';
 import './CreateCommunityGroupPage.scss';
 
 const CreateCommunityGroupPage = () => {
@@ -28,68 +28,7 @@ const CreateCommunityGroupPage = () => {
     <GroupForm
       formType="create"
       onSubmit={(values) => {
-        const payload = {
-          chainId: app.activeChainId(),
-          address: app.user.activeAccount.address,
-          groupName: values.groupName,
-          groupDescription: values.groupDescription,
-          topicIds: values.topics.map((x) => x.value),
-          requirementsToFulfill:
-            values.requirementsToFulfill === 'ALL'
-              ? values.requirements.length
-              : values.requirementsToFulfill,
-          requirements: [],
-        };
-
-        // map requirements and add to payload
-        values.requirements.map((x) => {
-          if (
-            x.requirementType === SPECIFICATIONS.ERC_20 ||
-            x.requirementType === SPECIFICATIONS.ERC_721
-          ) {
-            payload.requirements.push({
-              rule: 'threshold',
-              data: {
-                threshold: x.requirementAmount,
-                source: {
-                  source_type: x.requirementType,
-                  evm_chain_id: parseInt(x.requirementChain),
-                  contract_address: x.requirementContractAddress,
-                },
-              },
-            });
-            return;
-          }
-
-          if (x.requirementType === TOKENS.COSMOS_TOKEN) {
-            payload.requirements.push({
-              rule: 'threshold',
-              data: {
-                threshold: x.requirementAmount,
-                source: {
-                  source_type: x.requirementType,
-                  cosmos_chain_id: x.requirementChain,
-                  token_symbol: 'COS',
-                },
-              },
-            });
-            return;
-          }
-
-          if (x.requirementType === TOKENS.EVM_TOKEN) {
-            payload.requirements.push({
-              rule: 'threshold',
-              data: {
-                threshold: x.requirementAmount,
-                source: {
-                  source_type: x.requirementType,
-                  evm_chain_id: parseInt(x.requirementChain),
-                },
-              },
-            });
-            return;
-          }
-        });
+        const payload = makeGroupDataBaseAPIPayload(values);
 
         createGroup(payload)
           .then(() => {
